backend/requirements.txt
fastapi
uvicorn[standard]
pandas
pybaseball
numpy
python-dateutil
python-multipart
requests
